requirements.txt
Django==5.2.11
django-jazzmin==3.0.4
python-dotenv==1.0.0
djangorestframework==3.17.1
django-filter==25.1
django-crispy-forms==2.4
crispy-bootstrap5==2026.3
django-widget-tweaks==1.5.0
Markdown==3.8
gunicorn==23.0.0
psycopg2-binary==2.9.10
dj-database-url==2.3.0
celery==5.4.0
redis==5.2.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.2.0
requests==2.32.3
python-decouple==3.8
whitenoise==6.9.0
bleach==6.0.0
django-axes==6.1.0
django-ratelimit==4.0.0
drf_yasg==1.21.15
django-countries==7.6.1
